:add: dynamic whitelist page
