+++ components/admin/CreateWhitelistEventForm.jsx
@@ -163,7 +163,6 @@ const CreateWhitelistEventForm = () => {
       let updateErrors = {};
       Object.entries(formData).forEach((entry) => {
         const [key, value] = entry;
-        console.log([key, value]);
         if (!value && Object.hasOwnProperty.call(formErrors, key)) {
           let newEntry = { [key]: true };
           updateErrors = { ...updateErrors, ...newEntry };
