+++ components/admin/EditWhitelistEventForm.jsx
@@ -217,7 +217,6 @@ const EditWhitelistEventForm = () => {
       let updateErrors = {};
       Object.entries(formData).forEach((entry) => {
         const [key, value] = entry;
-        console.log([key, value]);
         if (!value && Object.hasOwnProperty.call(formErrors, key)) {
           let newEntry = { [key]: true };
           updateErrors = { ...updateErrors, ...newEntry };
